feat(filters): make filters layout a sidebar
