+++ src/components/RecommendationFilters/RecommendationFilters.module.css
@@ -1,17 +1,9 @@
 /* Recommendation Filters Component Styles */
 
 .filtersContainer {
-  background: var(--surface);
-  border: 1px solid var(--border);
-  border-radius: 12px;
-  padding: 1rem;
-  margin-bottom: 1.5rem;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  transition: all 0.2s ease;
-}
-
-.filtersContainer:hover {
-  box-shadow: 0 4px 16px rgba(0, 0, 0, 0.15);
+  /* Container styling is now handled by parent sidebar */
+  padding: 0;
+  margin: 0;
 }
 
 /* Filter Header */
@@ -157,95 +149,6 @@
   border-top: 1px solid var(--border);
 }
 
-/* Search Section */
-.searchSection {
-  display: flex;
-  flex-direction: column;
-  gap: 0.75rem;
-}
-
-.searchInput {
-  position: relative;
-  display: flex;
-  align-items: center;
-}
-
-.searchIcon {
-  position: absolute;
-  left: 0.75rem;
-  color: var(--text-secondary);
-  pointer-events: none;
-  z-index: 1;
-}
-
-.searchField {
-  width: 100%;
-  padding: 0.75rem 0.75rem 0.75rem 2.5rem;
-  border-radius: 8px;
-  background: var(--surface);
-  color: var(--text-primary);
-  font-size: 0.875rem;
-  transition: all 0.2s ease;
-}
-
-.searchField:focus {
-  outline: none;
-  border-color: var(--primary);
-  box-shadow: 0 0 0 3px rgba(var(--primary-rgb), 0.1);
-}
-
-.clearSearch {
-  position: absolute;
-  right: 0.5rem;
-  background: none;
-  border: none;
-  padding: 0.5rem;
-  cursor: pointer;
-  color: var(--text-secondary);
-  border-radius: 4px;
-  transition: all 0.2s ease;
-}
-
-.clearSearch:hover {
-  background: var(--surface-hover);
-  color: var(--text-primary);
-}
-
-/* Search Fields */
-.searchFields {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 0.75rem;
-}
-
-.searchFieldsLabel {
-  font-size: 0.75rem;
-  color: var(--text-secondary);
-  font-weight: 500;
-}
-
-.searchFieldOption {
-  display: flex;
-  align-items: center;
-  gap: 0.375rem;
-  font-size: 0.75rem;
-  color: var(--text-secondary);
-  cursor: pointer;
-  padding: 0.25rem 0.5rem;
-  border-radius: 4px;
-  transition: all 0.2s ease;
-}
-
-.searchFieldOption:hover {
-  background: var(--surface-hover);
-  color: var(--text-primary);
-}
-
-.searchFieldOption input[type="checkbox"] {
-  margin: 0;
-  cursor: pointer;
-}
 
 /* Section Titles */
 .sectionTitle {
@@ -262,11 +165,12 @@
   display: flex;
   flex-direction: column;
   gap: 0.75rem;
+  margin-bottom: var(--spacing-md);
 }
 
 .skillCategories {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+  display: flex;
+  flex-direction: column;
   gap: 0.5rem;
 }
 
@@ -448,7 +352,7 @@
   }
   
   .skillCategories {
-    grid-template-columns: 1fr;
+    flex-direction: column;
   }
   
   .sortOptions {
