+++ src/components/SearchBar/SearchBar.module.css
@@ -1,0 +1,157 @@
+.searchContainer {
+  background: var(--color-background);
+  border: 1px solid var(--color-border);
+  border-radius: var(--border-radius-lg);
+  padding: var(--spacing-lg);
+  box-shadow: var(--shadow-sm);
+  width: 100%;
+}
+
+.searchHeader {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: var(--spacing-lg);
+}
+
+.searchTitle {
+  display: flex;
+  align-items: center;
+  gap: var(--spacing-xs);
+  font-size: var(--font-size-base);
+  font-weight: 600;
+  color: var(--color-text);
+}
+
+.searchInput {
+  position: relative;
+  margin-bottom: var(--spacing-md);
+}
+
+.searchField {
+  width: 100%;
+  padding: var(--spacing-sm) var(--spacing-md);
+  padding-left: 36px;
+  border: 1px solid var(--color-border);
+  border-radius: var(--border-radius-md);
+  font-size: var(--font-size-base);
+  background: var(--color-background);
+  color: var(--color-text);
+  transition: all 0.2s ease;
+}
+
+.searchField:focus {
+  outline: none;
+  border-color: var(--color-primary);
+  box-shadow: 0 0 0 2px rgba(99, 102, 241, 0.1);
+}
+
+.searchField::placeholder {
+  color: var(--color-text-secondary);
+}
+
+.searchIcon {
+  position: absolute;
+  left: var(--spacing-sm);
+  top: 50%;
+  transform: translateY(-50%);
+  color: var(--color-text-secondary);
+  pointer-events: none;
+}
+
+.clearSearch {
+  position: absolute;
+  right: var(--spacing-sm);
+  top: 50%;
+  transform: translateY(-50%);
+  background: none;
+  border: none;
+  color: var(--color-text-secondary);
+  cursor: pointer;
+  padding: 2px;
+  border-radius: var(--border-radius-sm);
+  transition: all 0.2s ease;
+}
+
+.clearSearch:hover {
+  color: var(--color-text);
+  background: var(--color-surface);
+}
+
+.searchFields {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacing-sm);
+}
+
+.searchFieldsLabel {
+  font-size: var(--font-size-sm);
+  font-weight: 500;
+  color: var(--color-text-secondary);
+}
+
+.searchFieldsGrid {
+  display: flex;
+  flex-wrap: nowrap;
+  gap: var(--spacing-md);
+  overflow-x: auto;
+}
+
+.searchFieldOption {
+  display: flex;
+  align-items: center;
+  gap: var(--spacing-xs);
+  font-size: var(--font-size-sm);
+  color: var(--color-text-secondary);
+  cursor: pointer;
+  padding: var(--spacing-xs);
+  border-radius: var(--border-radius-sm);
+  transition: all 0.2s ease;
+  white-space: nowrap;
+  flex-shrink: 0;
+}
+
+.searchFieldOption:hover {
+  background: var(--color-surface);
+  color: var(--color-text);
+}
+
+.searchFieldOption input[type="checkbox"] {
+  margin: 0;
+  cursor: pointer;
+}
+
+.searchFieldOption span {
+  user-select: none;
+}
+
+/* Responsive Design */
+@media (max-width: 1024px) {
+  .searchFieldsGrid {
+    grid-template-columns: repeat(auto-fit, minmax(110px, 1fr));
+  }
+}
+
+@media (max-width: 768px) {
+  .searchContainer {
+    padding: var(--spacing-md);
+  }
+
+  .searchHeader {
+    margin-bottom: var(--spacing-md);
+  }
+
+  .searchFieldsGrid {
+    grid-template-columns: repeat(auto-fit, minmax(90px, 1fr));
+    gap: var(--spacing-xs);
+  }
+
+  .searchFieldOption {
+    padding: var(--spacing-xs);
+    font-size: var(--font-size-xs);
+  }
+
+  .searchFields {
+    gap: var(--spacing-xs);
+  }
+}
