+++ src/components/SortingControls/SortingControls.module.css
@@ -1,10 +1,11 @@
 .sortingCard {
   background: var(--color-background);
   border: 1px solid var(--color-border);
-  border-radius: var(--border-radius-md);
-  padding: var(--spacing-md);
+  border-radius: var(--border-radius-lg);
+  padding: var(--spacing-lg);
   margin-bottom: var(--spacing-lg);
   box-shadow: var(--shadow-sm);
+  width: 100%;
 }
 
 .sortingHeader {
